move and rename draw::draw_rgb to utils::blit_rgb

+++ fltk/src/draw/routines.rs
@@ -1,4 +1,4 @@
-use super::types::{Coord, Coordf};
+use super::types::{Coord, Coordf, Rect};
 use crate::enums::{Align, Color, ColorDepth, Cursor, Font, FrameType, Shortcut};
 use crate::image::RgbImage;
 use crate::prelude::*;
@@ -225,9 +225,9 @@ pub fn set_draw_color(color: Color) {
 }
 
 /// Draws a circle
-pub fn draw_circle(x: f64, y: f64, r: f64) {
+pub fn draw_circle(x: i32, y: i32, r: f64) {
     unsafe {
-        Fl_circle(x, y, r);
+        Fl_circle(x as f64, y as f64, r);
     }
 }
